refactor: extract shared mocks into mocks_test.go
