+++ internal/controllers/worktree/new_test.go
@@ -23,18 +23,6 @@ func (m *mockWorktreeCreator) CopyIgnoredFiles(mainPath, worktreePath string) []
 	return m.returnWarnings
 }
 
-type mockSessionLauncher struct {
-	returnErr  error
-	calledName string
-	calledDir  string
-}
-
-func (m *mockSessionLauncher) LaunchSession(name, dir string) error {
-	m.calledName = name
-	m.calledDir = dir
-	return m.returnErr
-}
-
 func TestNewControllerHandle(t *testing.T) {
 	t.Run("missingName", func(t *testing.T) {
 		controller := NewNewController(
